#14 [FIX] 지역코드 범위 넘어가서 Long으로 수정

+++ saju-backend/src/main/java/com/saju/sajubackend/api/matching/dto/MatchingMemberResponseDto.java
@@ -10,7 +10,7 @@ public record MatchingMemberResponseDto(
         String nickname,
         int score,
         String profileImage,
-        Integer region,
+        Long region,
         int age,
         String celestialStem,
         String introduction
