+++ saju-backend/src/main/java/com/saju/sajubackend/api/member/domain/Member.java
@@ -1,12 +1,5 @@
 package com.saju.sajubackend.api.member.domain;
 
-import com.saju.sajubackend.common.converter.DrinkingFrequencyConverter;
-import com.saju.sajubackend.common.converter.ReligionConverter;
-import com.saju.sajubackend.common.converter.SmokingStatusConverter;
-import com.saju.sajubackend.common.enums.CelestialStem;
-import com.saju.sajubackend.common.enums.DrinkingFrequency;
-import com.saju.sajubackend.common.enums.Religion;
-import com.saju.sajubackend.common.enums.SmokingStatus;
 import com.saju.sajubackend.common.converter.*;
 import com.saju.sajubackend.common.enums.*;
 import jakarta.persistence.*;
@@ -50,7 +43,7 @@ public class Member {
     private Integer height;
 
     @Column(name = "city_code")
-    private Integer cityCode;
+    private Long cityCode;
 
     @Column(name = "age")
     private Integer age;
